+++ geekshop/adminapp/templates/adminapp/book/book_delete.html
