+++ geekshop/adminapp/templates/adminapp/publisher/publisher_delete.html
