+++ geekshop/adminapp/templates/adminapp/publisher/publisher_update.html
