+++ geekshop/adminapp/templates/adminapp/translator/translator.html
