+++ geekshop/adminapp/templates/adminapp/translator/translator_delete.html
