+++ geekshop/adminapp/templates/adminapp/users/user_create.html
@@ -5,7 +5,7 @@
     <div class="col-3 mx-auto">
     <form class="form-horizontal" action="{% url 'adminapp:user_create' %}" method="post" enctype="multipart/form-data">
         {% csrf_token %}
-        {{ user_form.as_p }}
+        {{ form.as_p }}
 
         <input class="form-control" type="submit" value="Создать">
     </form>
