+++ geekshop/adminapp/templates/adminapp/users/user_delete.html
@@ -2,17 +2,17 @@
 {% load static %}
 
 
-{% block content %}
+{% block content %}  object
   <div class="user_delete">
-    Уверены, что хотите удалить {{ user_to_delete.username }}
-    {% if user_to_delete.first_name %}
-      <b>{{ user_to_delete.first_name|title }}</b>
-      {% if user_to_delete.last_name %}
-        {{ user_to_delete.last_name|title }}
+    Уверены, что хотите удалить {{ object.username }}
+    {% if object.first_name %}
+      <b>{{ object.first_name|title }}</b>
+      {% if object.last_name %}
+        {{ object.last_name|title }}
       {% endif %}
     {% endif %}
     ?
-    <form action="{% url 'adminapp:user_delete' user_to_delete.pk %}"
+    <form action="{% url 'adminapp:user_delete' object.pk %}"
           method="post">
       {% csrf_token %}
       <input class="btn btn-danger" type="submit" value="удалить">
