+++ geekshop/adminapp/templates/adminapp/users/user_update.html
@@ -4,7 +4,7 @@
 {% block content %}
   <form class="form-horizontal" method="post" enctype="multipart/form-data">
     {% csrf_token %}
-    {{ user_form.as_p }}
+    {{ form.as_p }}
     <input class="form-control" type="submit" value="сохранить">
   </form>
   <button class="btn btn-round form-control last">
