+++ geekshop/adminapp/templatetags/my_tags.py
@@ -1,0 +1,14 @@
+from django import template
+from django.conf import settings
+
+register = template.Library()
+
+
+@register.filter(name='media_folder_users')
+def media_folder_users(string):
+    if not string:
+        string = 'users_avatars/default.jpg'
+    return f'{settings.MEDIA_URL}{string}'
+
+
+register.filter('media_folder_users', media_folder_users)
