+++ geekshop/authnapp/templates/authnapp/edit.html
@@ -1,5 +1,6 @@
 {% extends 'authnapp/base.html' %}
 {% load static %}
+{% load my_tags %}
 
 
 {% block content %}
@@ -13,9 +14,5 @@
             на главную
         </a>
     </button>
-    {% if user.avatar %}
-    <div class="user_avatar"><img src="/media/{{ user.avatar }}"></div>
-    {% else %}
-        <p>Аватар отсутствует.</p>
-    {% endif %}
+    <div class="user_avatar"><img src="{{ user.avatar|media_folder_users }}" width="10%"></div>
 {% endblock %}
