+++ geekshop/mainapp/templates/mainapp/dynamic_catalog.html
@@ -3,43 +3,71 @@
 {% load static %}
 
 {% block title %}
-    <title>{{ title }}</title>
+<title>{{ title }}</title>
 {% endblock %}
 
 {% block style %}
-    <link rel="stylesheet" type="text/css" href="{% static 'mainapp/style/style.css' %}">
+<link rel="stylesheet" type="text/css" href="{% static 'mainapp/style/style.css' %}">
 {% endblock %}
 
 {% block content %}
-    <div class="header">
+<div class="header">
 
-        {% include 'logo.html' %}
+    {% include 'logo.html' %}
 
-        {% include 'basket_status.html' %}
+    {% include 'basket_status.html' %}
 
-        <div class="menu">
-            <h1><b>Меню:</b></h1>
-            <ul class="menu">
-                <li><a href="{% url 'index' %}">Главная</a><br></li>
-                <li><a href="{% url 'contacts' %}">Контакты</a><br></li>
-            </ul>
-        </div>
-
-        {% include 'sign.html' %}
-
+    <div class="menu">
+        <h1><b>Меню:</b></h1>
+        <ul class="menu">
+            <li><a href="{% url 'index' %}">Главная</a><br></li>
+            <li><a href="{% url 'contacts' %}">Контакты</a><br></li>
+        </ul>
     </div>
 
-    <hr>
+    {% include 'sign.html' %}
 
-    <h2 class="catalog">Каталог товаров:</h2>
+</div>
 
-    <div class="catalog-list">
-        {% for book in books %}
-            <div class="book">
-                <img src="{{ book.image.url }}" width="360" height="420"><br>
-                <a href="{% url 'products:book' id=book.book_id %}#main">{{ book.name }} - {{ book.author }}<a><br>
-            </div>
-        {% endfor %}
+<hr>
+
+<h2 class="catalog">Каталог товаров:</h2>
+
+<div class="catalog-list">
+    {% for book in books %}
+    <div class="book">
+        <img src="{{ book.image.url }}" width="360" height="420"><br>
+        <a href="{% url 'products:book' id=book.book_id %}#main">{{ book.name }} - {{ book.author }}<a><br>
     </div>
+    {% endfor %}
+</div>
+
+<hr>
+
+{% for book in page_obj %}
+    <div class="book">
+        <img src="{{ book.image.url }}" width="10%"><br>
+        <a href="{% url 'products:book' id=book.pk %}#main">{{ book.name }} - {{ book.author }}<a><br>
+
+    </div>
+{% endfor %}
+
+<div class="pagination">
+    <span class="step-links">
+        {% if page_obj.has_previous %}
+            <a href="?page=1">&laquo; Первая</a>
+            <a href="?page={{ page_obj.previous_page_number }}">Предыдущая</a>
+        {% endif %}
+
+        <span class="current">
+            Страница {{ page_obj.number }} из {{ page_obj.paginator.num_pages }}.
+        </span>
+
+        {% if page_obj.has_next %}
+            <a href="?page={{ page_obj.next_page_number }}">Следующая</a>
+            <a href="?page={{ page_obj.paginator.num_pages }}">Последняя &raquo;</a>
+        {% endif %}
+    </span>
+</div>
 
 {% endblock %}
